Calculion 0.0.1 release x 3.
This commit is the last in a commit chain officially publishing the
first release of this web app: Calculion 0.0.1. Specifically, this
commit revises the foreword of the default tab to explicitly link to
both the open-source GitHub repository underlying this web app and the
issue tracker hosted at that repository. Let's do this for the good of
calculating lions. (*Final finale!*)